requirements.txt
Django~=3.2.10
django-filter~=21.1
djangorestframework~=3.13.1
djangorestframework-simplejwt~=5.0.0
django-cors-headers~=3.10.1
drf-dynamic-fields~=0.3.1
drf-extensions~=0.7.1
mollie-api-python~=2.12.0